[FIX] account.invoice report: show customer description of payment term
The payment term description (note) is meant
to be shown to customers. Restore it in the
default invoice report, as it was in 7.0.
If empty, assume the payment term is meant
to be hidden from customer, and do not show
anything.
By default the description includes the name
of the payment term, so no need to keep both.