[REV] stock: revert of 75a2b9650b19b436cd24f9751a912a92a35bb1c5
Reintroduction of field mes_type on product.template. This field was not
used in the code anymore, but we keep it for future use.

This field allows the user to express that there is no fixed conversion
rate between uom and uos.

(also very partial revert of a1032c7d00fcecba337185995fba9da11046004d)