[FIX] sale_stock: product_id_change_with_wh on sale order popup too
product_id_change_with_wh onchange was not anymore called when using the popup form in sale order line.

The behavior was different concerning the product onchange when being in the editable list or in the popup form
For instance, for make to stock products not having enough stock, the warning was displayed when being in editable list, but not in the popup form

This is a regression of rev. ed1fabd6f00a720678cc690e5311a2019a49bd67

opw-619624