[FIX] hr_timesheet_sheet: prevent sheets overlaping
The constraint prevent sheets to overlap was broken because it relies on a check on the
user_id.  The problem is that the latter uses an old-api function field
that is not recomputed yet at the time of the validation.  The fix consists in
using a non-computed field instead.

This rev. is related to:
 - 178e73e8bb5a35ecc874266df9aa8910b77ec106
 - 8f51a6665868e26d99bafef884f76a37a41845c2

Closes #4217
opw-620175