[FIX] l10n_be_coda: globalisation stack
The first line having the globalisation code is the globalized amount,
then the next line having this same globalisation code
is the last line of the globalisation (the last children).

The first line must no be included in the bank statement,
but the last line must (as this is a real children line)

This is a regression of commit: cbc52f80eb0424a6bc2947d10d150dd3833d0100

Official doc concerning globalization code:
The value which is mentioned (1 to 9), specifies the hierarchy level of the
globalisation of which this record is the first.
The same code will be repeated at the end of the globalisation