[FIX] point_of_sale: share is not a dependence of point_of_sale
This reverts commit b0f114d385c19f07608cfa7980fae43737a1c4cc.

point_of_sale does not depends on share. This domain can therefore not be applied.
It works for new databases as the module share is auto installed.
But as soon as the module share is uninstalled, this domain will lead to a crash
There is no way to "solve" this issue, other than by a view customization
if the issue is critical for the customer.

I did not pay enough attention when I reviewed the PR.