[FIX] base: ir_config_parameter increase database.uuid.private entropy
Use uuid4 instead of uuid1 to get 128 bits of entropy.

Rename database.uuid.private to the more meaningful database.secret as it
explains the usage instead of the technical details to generate it. And we
might switch to a longer string with more entropy in the future.

This string can be used to sign cookies or token with
sign(src) -> (src, hash(src + secret))
check(msg) -> msg[1] == hash(msg[0] + secret)