[FIX] hr_timesheet_sheet: prevent modify activities of confirmed sheets
In 7.0, prevent changes in activities of validated sheets was done thanks to the constraint
_check_sheet_state.
In 7.0, python constraints were checked at each write operations
whatever if the fields on which the constraint is were altered or not

This is no longer the case in Odoo 8.0: The constraint is checked
only if the fields on which the constraint is are altered.

As this specific constraint must be applied anytime, whatever the altered fields,
we now do this constraint check in the "write" method.
Besides, it was already the case for the unlink method.

opw-627415