[FIX] web: context lang & active* in action buttons
This reverts rev. 84fe184b70fcb62a348828e85ab107733e40125a

As explained in 84fe184b70fcb62a348828e85ab107733e40125a,
if a field "lang" is present in the view, clicking in any action item
in the more menu leaded for the action title to be translated
in the lang value of the form, such as in the partner form.

84fe184b70fcb62a348828e85ab107733e40125a fixed the issue,
but has as side-effect to not update correctly the active* keys.
So, if "active_id" was used in the context of the action button,
and the active_id was set in the dataset context, for example
because you come from another form, trough another action button
(For instance, Customers > Opportunities > Logged Calls),
the active_id was not updated with the current id of the record.

opw-620293
opw-617321
fixes #3462